Add support for nginx reverse proxy
If nginx is used in a reverse proxy, REMOTE_ADDR will be set to the
local IP address which is useless for spam detection. The solution
to this is to use proxy_set_header X-Real-IP ; in nginx

to use that header, all of the code needed to set remote address to
['REMOTE_ADDR'] initally and then, if it was set, use
HTTP_X_REAL_IP to set the true IP.

This allows admins to see where spam is truly coming from once again.

For more info see the following link:
https://easyengine.io/tutorials/nginx/forwarding-visitors-real-ip/

Signed-off-by: Simon Sickle <simon@simonsickle.com>